# openapi/Penguin.yaml
openapi: 3.0.3
info:
  version: 0.6.0
  title: Penguin API
  description: "A list of the currently provided Interfaces to interact with the penguin, implementing file operations and sending messages"
security:
  - {}
...
servers:
  - url: "http://{apiRoot}:{port}/v1"
    variables:
      apiRoot:
        default: "localhost"
        description: Base address of the local pen node main API
      port:
        default: "1633"
        description: Service port provided in pen node config
  - url: "http://{apiRoot}:{port}"
    variables:
      apiRoot:
        default: "localhost"
        description: Base address of the local pen node main API
      port:
        default: "1633"
        description: Service port provided in pen node config
paths:
  "/bytes":
    post:
      summary: "Upload data"
      tags:
        - Bytes
      parameters:
        - $ref: "PenguinCommon.yaml#/components/parameters/PenguinTagParameter"
        - $ref: "PenguinCommon.yaml#/components/parameters/PenguinPinParameter"
        - $ref: "PenguinCommon.yaml#/components/parameters/PenguinEncryptParameter"
        - $ref: "PenguinCommon.yaml#/components/parameters/PenguinPostageBatchId"
      requestBody:
        content:
          application/octet-stream:
            schema:
              type: string
              format: binary
      responses:
        "201":
          description: Ok
          headers:
            "penguin-tag":
              $ref: "PenguinCommon.yaml#/components/headers/PenguinTag"
          content:
            application/json:
              schema:
                $ref: "PenguinCommon.yaml#/components/schemas/ReferenceResponse"
        "400":
          $ref: "PenguinCommon.yaml#/components/responses/400"
        "403":
          $ref: "PenguinCommon.yaml#/components/responses/403"
        "500":
          $ref: "PenguinCommon.yaml#/components/responses/500"
        default:
          description: Default response
  "/bytes/{reference}":
    get:
      summary: "Get referenced data"
      tags:
        - Bytes
      parameters:
        - in: path
          name: reference
          schema:
            $ref: "PenguinCommon.yaml#/components/schemas/PenguinReference"
          required: true
          description: Penguin address reference to content
      responses:
        "200":
          description: Retrieved content specified by reference
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        "404":
          $ref: "PenguinCommon.yaml#/components/responses/404"
        default:
          description: Default response
  "/chunks/{reference}":
    get:
      summary: "Get Chunk"
      tags:
        - Chunk
      parameters:
        - in: path
          name: reference
          schema:
            $ref: "PenguinCommon.yaml#/components/schemas/PenguinReference"
          required: true
          description: Penguin address of chunk
        - $ref: "PenguinCommon.yaml#/components/parameters/PenguinRecoveryTargetsParameter"
      responses:
        "200":
          description: Retrieved chunk content
          headers:
            "penguin-recovery-targets":
              $ref: "PenguinCommon.yaml#/components/headers/PenguinRecoveryTargets"
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        "202":
          description: chunk recovery initiated. retry after sometime.
        "400":
          $ref: "PenguinCommon.yaml#/components/responses/400"
        "404":
          $ref: "PenguinCommon.yaml#/components/responses/404"
        "500":
          $ref: "PenguinCommon.yaml#/components/responses/500"
        default:
          description: Default response
  "/chunks":
    post:
      summary: "Upload Chunk"
      tags:
        - Chunk
      parameters:
        - $ref: "PenguinCommon.yaml#/components/parameters/PenguinTagParameter"
        - $ref: "PenguinCommon.yaml#/components/parameters/PenguinPinParameter"
        - $ref: "PenguinCommon.yaml#/components/parameters/PenguinPostageBatchId"
      requestBody:
        content:
          application/octet-stream:
            schema:
              type: string
              format: binary
      responses:
        "201":
          description: Ok
          headers:
            "penguin-tag":
              $ref: "PenguinCommon.yaml#/components/headers/PenguinTag"
          content:
            application/json:
              schema:
                $ref: "PenguinCommon.yaml#/components/schemas/Status"
        "400":
          $ref: "PenguinCommon.yaml#/components/responses/400"
        "500":
          $ref: "PenguinCommon.yaml#/components/responses/500"
        default:
          description: Default response
  "/pen":
    post:
      summary: "Upload file or a collection of files"
      description: "In order to upload a collection, user can send a multipart request with all the files populated in the form data with appropriate headers.\n\n
        User can also upload a tar file along with the penguin-collection header. This will upload the tar file after extracting the entire directory structure.\n\n
        If the penguin-collection header is absent, all requests (including tar files) are considered as single file uploads.\n\n
        A multipart request is treated as a collection regardless of whether the penguin-collection header is present. This means in order to serve single files
        uploaded as a multipart request, the penguin-index-document header should be used with the name of the file."
      tags:
        - File
        - Collection
      parameters:
        - in: query
          name: name
          schema:
            $ref: "PenguinCommon.yaml#/components/schemas/FileName"
          required: false
          description: Filename when uploading single file
        - $ref: "PenguinCommon.yaml#/components/parameters/PenguinTagParameter"
        - $ref: "PenguinCommon.yaml#/components/parameters/PenguinPinParameter"
        - $ref: "PenguinCommon.yaml#/components/parameters/PenguinEncryptParameter"
        - $ref: "PenguinCommon.yaml#/components/parameters/ContentTypePreserved"
        - $ref: "PenguinCommon.yaml#/components/parameters/PenguinCollection"
        - $ref: "PenguinCommon.yaml#/components/parameters/PenguinIndexDocumentParameter"
        - $ref: "PenguinCommon.yaml#/components/parameters/PenguinErrorDocumentParameter"
        - $ref: "PenguinCommon.yaml#/components/parameters/PenguinPostageBatchId"
      requestBody:
        content:
          multipart/form-data:
            schema:
              properties:
                file:
                  type: array
                  items:
                    type: string
                    format: binary
          application/octet-stream:
            schema:
              type: string
              format: binary
          application/x-tar:
            schema:
              type: string
              format: binary
      responses:
        "201":
          description: Ok
          headers:
            "penguin-tag":
              $ref: "PenguinCommon.yaml#/components/headers/PenguinTag"
            "etag":
              $ref: "PenguinCommon.yaml#/components/headers/ETag"
          content:
            application/json:
              schema:
                $ref: "PenguinCommon.yaml#/components/schemas/ReferenceResponse"
        "400":
          $ref: "PenguinCommon.yaml#/components/responses/400"
        "403":
          $ref: "PenguinCommon.yaml#/components/responses/403"
        "500":
          $ref: "PenguinCommon.yaml#/components/responses/500"
        default:
          description: Default response
  "/pen/{reference}":
    patch:
      summary: "Reupload a root hash to the network"
      tags:
        - Reupload file
        - Reupload collection
      parameters:
        - in: path
          name: reference
          schema:
            $ref: "PenguinCommon.yaml#/components/schemas/PenguinReference"
          required: true
          description: Root hash of content
      responses:
        "200":
          description: Ok
        "500":
          $ref: "PenguinCommon.yaml#/components/responses/500"
        default:
          description: Default response
    get:
      summary: "Get file or index document from a collection of files"
      tags:
        - Collection
      parameters:
        - in: path
          name: reference
          schema:
            $ref: "PenguinCommon.yaml#/components/schemas/PenguinReference"
          required: true
          description: Penguin address of content
        - $ref: "PenguinCommon.yaml#/components/parameters/PenguinRecoveryTargetsParameter"
      responses:
        "200":
          description: Ok
          headers:
            "penguin-recovery-targets":
              $ref: "PenguinCommon.yaml#/components/headers/PenguinRecoveryTargets"
          # "penguin-feed-index":
          #   $ref: "PenguinCommon.yaml#/components/headers/PenguinFeedIndex"
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        "400":
          $ref: "PenguinCommon.yaml#/components/responses/400"
        "404":
          $ref: "PenguinCommon.yaml#/components/responses/404"
        "500":
          $ref: "PenguinCommon.yaml#/components/responses/500"
        default:
          description: Default response
  "/pen/{reference}/{path}":
    get:
      summary: "Get referenced file from a collection of files"
      tags:
        - Collection
      parameters:
        - in: path
          name: reference
          schema:
            $ref: "PenguinCommon.yaml#/components/schemas/PenguinReference"
          required: true
          description: Penguin address of content
        - in: path
          name: path
          schema:
            type: string
          required: true
          description: Path to the file in the collection.
        - $ref: "PenguinCommon.yaml#/components/parameters/PenguinRecoveryTargetsParameter"
      responses:
        "200":
          description: Ok
          headers:
            "penguin-recovery-targets":
              $ref: "PenguinCommon.yaml#/components/headers/PenguinRecoveryTargets"
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary

        "400":
          $ref: "PenguinCommon.yaml#/components/responses/400"
        "404":
          $ref: "PenguinCommon.yaml#/components/responses/404"
        "500":
          $ref: "PenguinCommon.yaml#/components/responses/500"
        default:
          description: Default response
  "/tags":
    get:
      summary: Get list of tags
      tags:
        - Tag
      parameters:
        - in: query
          name: offset
          schema:
            type: integer
            minimum: 0
            default: 0
          required: false
          description: The number of items to skip before starting to collect the result set.
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 1000
            default: 100
          required: false
          description: The numbers of items to return.
      responses:
        "200":
          description: List of tags
          content:
            application/json:
              schema:
                $ref: "PenguinCommon.yaml#/components/schemas/TagsList"
        "403":
          $ref: "PenguinCommon.yaml#/components/responses/403"
        "500":
          $ref: "PenguinCommon.yaml#/components/responses/500"
        default:
          description: Default response
    post:
      summary: "Create Tag"
      tags:
        - Tag
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "PenguinCommon.yaml#/components/schemas/NewTagRequest"
      responses:
        "201":
          description: New Tag Info
          content:
            application/json:
              schema:
                $ref: "PenguinCommon.yaml#/components/schemas/NewTagResponse"
        "403":
          $ref: "PenguinCommon.yaml#/components/responses/403"
        "500":
          $ref: "PenguinCommon.yaml#/components/responses/500"
        default:
          description: Default response
  "/tags/{uid}":
    get:
      summary: "Get Tag information using Uid"
      tags:
        - Tag
      parameters:
        - in: path
          name: uid
          schema:
            $ref: "PenguinCommon.yaml#/components/schemas/Uid"
          required: true
          description: Uid
      responses:
        "200":
          description: Tag info
          content:
            application/json:
              schema:
                $ref: "PenguinCommon.yaml#/components/schemas/NewTagResponse"
        "400":
          $ref: "PenguinCommon.yaml#/components/responses/400"
        "403":
          $ref: "PenguinCommon.yaml#/components/responses/403"
        "404":
          $ref: "PenguinCommon.yaml#/components/responses/404"
        "500":
          $ref: "PenguinCommon.yaml#/components/responses/500"
        default:
          description: Default response
    delete:
      summary: "Delete Tag information using Uid"
      tags:
        - Tag
      parameters:
        - in: path
          name: uid
          schema:
            $ref: "PenguinCommon.yaml#/components/schemas/Uid"
          required: true
          description: Uid
      responses:
        "204":
          $ref: "PenguinCommon.yaml#/components/responses/204"
        "400":
          $ref: "PenguinCommon.yaml#/components/responses/400"
        "403":
          $ref: "PenguinCommon.yaml#/components/responses/403"
        "404":
          $ref: "PenguinCommon.yaml#/components/responses/404"
        "500":
          $ref: "PenguinCommon.yaml#/components/responses/500"
        default:
          description: Default response
    patch:
      summary: "Update Total Count and penguin hash for a tag of an input stream of unknown size using Uid"
      tags:
        - Tag
      parameters:
        - in: path
          name: uid
          schema:
            $ref: "PenguinCommon.yaml#/components/schemas/Uid"
          required: true
          description: Uid
      requestBody:
        description: Can contain penguin hash to use for the tag
        required: false
        content:
          application/json:
            schema:
              $ref: "PenguinCommon.yaml#/components/schemas/Address"
      responses:
        "200":
          description: Ok
          content:
            application/json:
              schema:
                $ref: "PenguinCommon.yaml#/components/schemas/Status"
        "403":
          $ref: "PenguinCommon.yaml#/components/responses/403"
        "404":
          $ref: "PenguinCommon.yaml#/components/responses/404"
        "500":
          $ref: "PenguinCommon.yaml#/components/responses/500"
        default:
          description: Default response
  "/pins/{reference}":
    parameters:
      - in: path
        name: reference
        schema:
          $ref: "PenguinCommon.yaml#/components/schemas/PenguinOnlyReference"
        required: true
        description: Penguin reference of the root hash
    post:
      summary: Pin the root hash with the given reference
      tags:
        - Root hash pinning
      responses:
        "200":
          description: Pin already exists, so no operation
          content:
            application/json:
              schema:
                $ref: "PenguinCommon.yaml#/components/schemas/Response"
        "201":
          description: New pin with root reference was created
          content:
            application/json:
              schema:
                $ref: "PenguinCommon.yaml#/components/schemas/Response"
        "400":
          $ref: "PenguinCommon.yaml#/components/responses/400"
        "403":
          $ref: "PenguinCommon.yaml#/components/responses/403"
        "404":
          $ref: "PenguinCommon.yaml#/components/responses/404"
        "500":
          $ref: "PenguinCommon.yaml#/components/responses/500"
        default:
          description: Default response
    delete:
      summary: Unpin the root hash with the given reference
      tags:
        - Root hash pinning
      responses:
        "200":
          description: Unpinning root hash with reference
          content:
            application/json:
              schema:
                $ref: "PenguinCommon.yaml#/components/schemas/Response"
        "400":
          $ref: "PenguinCommon.yaml#/components/responses/400"
        "403":
          $ref: "PenguinCommon.yaml#/components/responses/403"
        "500":
          $ref: "PenguinCommon.yaml#/components/responses/500"
        default:
          description: Default response
    get:
      summary: Get pinning status of the root hash with the given reference
      tags:
        - Root hash pinning
      responses:
        "200":
          description: Reference of the pinned root hash
          content:
            application/json:
              schema:
                $ref: "PenguinCommon.yaml#/components/schemas/PenguinOnlyReference"
        "400":
          $ref: "PenguinCommon.yaml#/components/responses/400"
        "403":
          $ref: "PenguinCommon.yaml#/components/responses/403"
        "404":
          $ref: "PenguinCommon.yaml#/components/responses/404"
        "500":
          $ref: "PenguinCommon.yaml#/components/responses/500"
        default:
          description: Default response
  "/pins":
    get:
      summary: Get the list of pinned root hash references
      tags:
        - Root hash pinning
      responses:
        "200":
          description: List of pinned root hash references
          content:
            application/json:
              schema:
                $ref: "PenguinCommon.yaml#/components/schemas/PenguinOnlyReferencesList"
        "403":
          $ref: "PenguinCommon.yaml#/components/responses/403"
        "500":
          $ref: "PenguinCommon.yaml#/components/responses/500"
        default:
          description: Default response
  "/pss/send/{topic}/{targets}":
    post:
      summary: Send to recipient or target with Postal Service for Penguin
      tags:
        - Postal Service for Penguin
      parameters:
        - in: path
          name: topic
          schema:
            $ref: "PenguinCommon.yaml#/components/schemas/PssTopic"
          required: true
          description: Topic name
        - in: path
          name: targets
          schema:
            $ref: "PenguinCommon.yaml#/components/schemas/PssTargets"
          required: true
          description: Target message address prefix. If multiple targets are specified, only one would be matched.
        - in: query
          name: recipient
          schema:
            $ref: "PenguinCommon.yaml#/components/schemas/PssRecipient"
          required: false
          description: Recipient publickey
        - $ref: "PenguinCommon.yaml#/components/parameters/PenguinPostageBatchId"
      responses:
        "201":
          description: Subscribed to topic
        "400":
          $ref: "PenguinCommon.yaml#/components/responses/400"
        "500":
          $ref: "PenguinCommon.yaml#/components/responses/500"
        default:
          description: Default response
  "/pss/subscribe/{topic}":
    get:
      summary: Subscribe for messages on the given topic.
      tags:
        - Postal Service for Penguin
      parameters:
        - in: path
          name: topic
          schema:
            $ref: "PenguinCommon.yaml#/components/schemas/PssTopic"
          required: true
          description: Topic name
      responses:
        "200":
          description: Returns a WebSocket with a subscription for incoming message data on the requested topic.
        "500":
          $ref: "PenguinCommon.yaml#/components/responses/500"
        default:
          description: Default response
  "/soc/{owner}/{id}":
    post:
      summary: Upload single owner chunk
      tags:
        - Single owner chunk
      parameters:
        - in: path
          name: owner
          schema:
            $ref: "PenguinCommon.yaml#/components/schemas/EthereumAddress"
          required: true
          description: Owner
        - in: path
          name: id
          schema:
            $ref: "PenguinCommon.yaml#/components/schemas/HexString"
          required: true
          description: Id
        - in: query
          name: sig
          schema:
            $ref: "PenguinCommon.yaml#/components/schemas/HexString"
          required: true
          description: Signature
        - $ref: "PenguinCommon.yaml#/components/parameters/PenguinPinParameter"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "PenguinCommon.yaml#/components/schemas/ReferenceResponse"
        "400":
          $ref: "PenguinCommon.yaml#/components/responses/400"
        "401":
          $ref: "PenguinCommon.yaml#/components/responses/401"
        "500":
          $ref: "PenguinCommon.yaml#/components/responses/500"
        default:
          description: Default response
  "/feeds/{owner}/{topic}":
    post:
      summary: Create an initial feed root manifest
      tags:
        - Feed
      parameters:
        - in: path
          name: owner
          schema:
            $ref: "PenguinCommon.yaml#/components/schemas/EthereumAddress"
          required: true
          description: Owner
        - in: path
          name: topic
          schema:
            $ref: "PenguinCommon.yaml#/components/schemas/HexString"
          required: true
          description: Topic
        - in: query
          name: type
          schema:
            $ref: "PenguinCommon.yaml#/components/schemas/FeedType"
          required: false
          description: "Feed indexing scheme (default: sequence)"
        - $ref: "PenguinCommon.yaml#/components/parameters/PenguinPinParameter"
        - $ref: "PenguinCommon.yaml#/components/parameters/PenguinPostageBatchId"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "PenguinCommon.yaml#/components/schemas/ReferenceResponse"
        "400":
          $ref: "PenguinCommon.yaml#/components/responses/400"
        "401":
          $ref: "PenguinCommon.yaml#/components/responses/401"
        "500":
          $ref: "PenguinCommon.yaml#/components/responses/500"
        default:
          description: Default response
    get:
      summary: Find feed update
      tags:
        - Feed
      parameters:
        - in: path
          name: owner
          schema:
            $ref: "PenguinCommon.yaml#/components/schemas/EthereumAddress"
          required: true
          description: Owner
        - in: path
          name: topic
          schema:
            $ref: "PenguinCommon.yaml#/components/schemas/HexString"
          required: true
          description: Topic
        - in: query
          name: at
          schema:
            type: integer
          required: false
          description: "Timestamp of the update (default: now)"
        - in: query
          name: type
          schema:
            $ref: "PenguinCommon.yaml#/components/schemas/FeedType"
          required: false
          description: "Feed indexing scheme (default: sequence)"
      responses:
        "200":
          description: Latest feed update
          headers:
            "penguin-feed-index":
              $ref: "PenguinCommon.yaml#/components/headers/PenguinFeedIndex"
            "penguin-feed-index-next":
              $ref: "PenguinCommon.yaml#/components/headers/PenguinFeedIndexNext"
          content:
            application/json:
              schema:
                $ref: "PenguinCommon.yaml#/components/schemas/ReferenceResponse"
        "400":
          $ref: "PenguinCommon.yaml#/components/responses/400"
        "401":
          $ref: "PenguinCommon.yaml#/components/responses/401"
        "500":
          $ref: "PenguinCommon.yaml#/components/responses/500"
        default:
          description: Default response
  "/stamps":
    get:
      summary: Get all available stamps for this node
      tags:
        - Postage Stamps
      responses:
        "200":
          description: Returns an array of all available postage batches.
          content:
            application/json:
              schema:
                $ref: "PenguinCommon.yaml#/components/schemas/PostageBatchesResponse"

        default:
          description: Default response
  "/stamps/{id}":
    parameters:
      - in: path
        name: id
        schema:
          $ref: "PenguinCommon.yaml#/components/schemas/BatchID"
        required: true
        description: Penguin address of the stamp
    get:
      summary: Get an individual postage batch status
      tags:
        - Postage Stamps
      responses:
        "200":
          description: Returns an individual postage batch state
          content:
            application/json:
              schema:
                $ref: "PenguinCommon.yaml#/components/schemas/PostageBatch"
        "400":
          $ref: "PenguinCommon.yaml#/components/responses/400"
        default:
          description: Default response
  "/stamps/{amount}/{depth}":
    post:
      summary: Buy a new postage batch. Be aware, this endpoint create an on-chain transactions and transfers PEN from the node's Ethereum account and hence directly manipulates the wallet balance!
      tags:
        - Postage Stamps
      parameters:
        - in: path
          name: amount
          schema:
            type: integer
          required: true
          description: Amount of PEN added that the postage batch will have.
        - in: path
          name: depth
          schema:
            type: integer
          required: true
          description: Batch depth which specifies how many chunks can be signed with the batch. It is a logarithm. Must be higher than default bucket depth (16)
        - in: query
          name: label
          schema:
            type: string
          required: false
          description: An optional label for this batch
        - $ref: "PenguinCommon.yaml#/components/parameters/GasPriceParameter"
      responses:
        "201":
          description: Returns the newly created postage batch ID
          content:
            application/json:
              schema:
                $ref: "PenguinCommon.yaml#/components/schemas/BatchIDResponse"
        "400":
          $ref: "PenguinCommon.yaml#/components/responses/400"
        "500":
          $ref: "PenguinCommon.yaml#/components/responses/500"
        default:
          description: Default response